refactor: separate builder and setup into modules
Closes #27
